build: Add a script (and make target) for creating a new release
This replaces the old `mkchangelog.py` script, which required manually
copying around the changelog text. The new script automatically creates
a tag (with an interactive editing session), and copies the edited text
to the clipboard using OSC 52 afterwards.

This script does not (yet) automate pushing the created tag to the
remote origin, to allow inspecting it before it gets published.